# Conflicts:
#	core/src/com/unciv/ui/worldscreen/WorldScreen.kt

+++ core/src/com/unciv/ui/pickerscreens/ModManagementScreen.kt
@@ -22,10 +22,12 @@ import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.pickerscreens.ModManagementOptions.SortType
 import com.unciv.ui.popup.Popup
 import com.unciv.ui.popup.ToastPopup
-import com.unciv.ui.popup.YesNoPopup
+import com.unciv.ui.popup.ConfirmPopup
 import com.unciv.ui.utils.AutoScrollPane
+import com.unciv.ui.utils.BaseScreen
 import com.unciv.ui.utils.ExpanderTab
 import com.unciv.ui.utils.KeyCharAndCode
+import com.unciv.ui.utils.RecreateOnResize
 import com.unciv.ui.utils.WrappableLabel
 import com.unciv.ui.utils.extensions.UncivDateFormat.formatDate
 import com.unciv.ui.utils.extensions.UncivDateFormat.parseDate
@@ -54,7 +56,7 @@ import kotlin.math.max
 class ModManagementScreen(
     previousInstalledMods: HashMap<String, ModUIData>? = null,
     previousOnlineMods: HashMap<String, ModUIData>? = null
-): PickerScreen(disableScroll = true) {
+): PickerScreen(disableScroll = true), RecreateOnResize {
 
     private val modTable = Table().apply { defaults().pad(10f) }
     private val scrollInstalledMods = AutoScrollPane(modTable)
@@ -102,7 +104,7 @@ class ModManagementScreen(
             if (game.settings.tileSet !in tileSets) {
                 game.settings.tileSet = tileSets.first()
             }
-            game.setScreen(MainMenuScreen())
+            game.popScreen()
         }
         closeButton.keyShortcuts.add(KeyCharAndCode.BACK)
 
@@ -527,21 +529,23 @@ class ModManagementScreen(
     private fun installedButtonAction(mod: ModUIData) {
         syncInstalledSelected(mod.name, mod.button)
         refreshInstalledModActions(mod.ruleset!!)
-        rightSideButton.setText("Delete [${mod.name}]".tr())
+        val deleteText = "Delete [${mod.name}]"
+        rightSideButton.setText(deleteText.tr())
         // Don't let the player think he can delete Vanilla and G&K rulesets
         rightSideButton.isEnabled = mod.ruleset.folderLocation!=null
         showModDescription(mod.name)
         rightSideButton.clearListeners()
         rightSideButton.onClick {
             rightSideButton.isEnabled = false
-            YesNoPopup(
+            ConfirmPopup(
+                screen = this,
                 question = "Are you SURE you want to delete this mod?",
+                confirmText = deleteText,
                 action = {
                     deleteMod(mod.ruleset)
                     modActionTable.clear()
                     rightSideButton.setText("[${mod.name}] was deleted.".tr())
                 },
-                screen = this,
                 restoreDefault = { rightSideButton.isEnabled = true }
             ).open()
         }
@@ -590,11 +594,7 @@ class ModManagementScreen(
         modDescriptionLabel.setText(onlineModDescription + separator + installedModDescription)
     }
 
-    override fun resize(width: Int, height: Int) {
-        if (stage.viewport.screenWidth != width || stage.viewport.screenHeight != height) {
-            game.setScreen(ModManagementScreen(installedModInfo, onlineModInfo))
-        }
-    }
+    override fun recreate(): BaseScreen = ModManagementScreen(installedModInfo, onlineModInfo)
 
     companion object {
         val modsToHideAsUrl by lazy {
